requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
websockets>=10.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
pandas>=1.3.0
numpy>=1.21.0
python-dotenv>=0.19.1
httpx>=0.18.0
aiohttp>=3.8.0
requests>=2.26.0